Remove GetRequesterCount
From what I've observed, if a player joins the game after a spray has been
placed with the plugin, neither the plugin nor normal sprays will show up
for that player if they haven't already downloaded the spray file.
Checking to see if a new player has downloaded the spray is pointless
and just makes the plugin unresponsive for no benefit.